AdmissionsSimulation

AdmissionsSimulation is designed to simulate outcomes of graduate admissions, with the intent of helping make decisions about both initial offers and wait-list offers.

This package focuses on those students to whom offers of admission have been extended, and uses past applicants as proxies for current applicants to make predictions about whether they'll accept the offer. Because two "similar" students might end up making different final decisions, the recommended practice is to identify many different potential proxies and use the distribution of their decisions to simulate future outcomes.

The inputs to this process are:

records on previous applicants

criteria for deciding the similarity between two students, resulting in a matching function computing a value between 0 (no match) and 1 (a perfect match).

Applicant records

Records on previous applicants are in two forms: very general information is stored in program_history, a dictionary recording just a few bits of information for each program. For example:

    program_history = Dict(ProgramKey(season=2021, program="NS") => ProgramData(slots=15, napplicants=302, firstofferdate=Date("2021-01-13"), lastdecisiondate=Date("2021-04-15")),
                           ProgramKey(season=2021, program="CB") => ProgramData(slots=5,  napplicants=160, firstofferdate=Date("2021-01-6"),  lastdecisiondate=Date("2021-04-15")))

suffices to record aggregate data for two programs, "NS" and "CB", during the 2021 season (corresponding to a decision deadline of April 15, 2021). This records the target number of matriculants (slots), the total number of applications received, the date of the very first offers extended, and the date on which a decision was due.

Valid choices for program names are listed in AdmissionsSimulation.program_lookups; internally the code always uses the abbreviation, but the

Detailed applicant records only need to include applicants to whom an offer of admission was extended. The requirements are described by NormalizedApplicant(applicant; program_history).

You can load both the program history and data on applicants using read_program_history and read_applicant_data.

Match criteria

Applicants 1 and 2 are matched by the following function:

\[\exp\left( - \frac{(r_1 - r_2)^2}{2 \sigma_r^2} - \frac{(t_1 - t_2)^2}{2 \sigma_t^2}\right)\]

where $r$ refers to the normalized ranks and $t$ to the normalized offer date. The $\sigma$ parameters measure the standard deviation, i.e., the tolerance for mismatch (larger $\sigma$ are more tolerant of mismatch). Smaller $\sigma_r$ increases the importance of choosing applicants of similar rank, and would indicate that there may be a strong rank-dependent element to recruitment (e.g., "more competitive applicants are harder to recruit"). Smaller $\sigma_t$ increases the importance of the timing of the offer, and would indicate that wait-list offers should be treated quite differently from initial offers. The units of both $\sigma$ parameters are those of the NormalizedApplicants, i.e., with values ranging between 0 and 1. For example, setting $\sigma_t = 0.2$ would, in essence, break the decision period (typically mid-January to April 15th) into roughly 5 periods, and match primarily against applicants who were extended offers during the same period. Note that the applicant ranks will typically be quite heavily weighted to low values (e.g., a program that only accepts the top 20% of applicants will only exhibit ranks between 0.0 and 0.2), and as a consequence $\sigma_r$ must be smaller than this span to have large effect.

This fundamental matching function is augmented by options to:

require that the matching applicants are from the same program

exclude past applicants who had already rendered their decision by this point in the admissions season. (The latter is intended to support modeling the currently-undecided applicants solely in terms of prior applicants who were also undecided at this point.)

See match_function for specific details.

Analysis and simulations

Once you've entered student records and defined a matching function, then for each outstanding offer you can compute the match likelihood of previous applicants using match_likelihood. The sum of the returned list is a rough measure of the "number" of prior applicants deemed to be a good match for the applicant you are modeling; if this value is small, your criteria for matching may be too stringent. Conversely, if this value is approximately equal to the total number of prior applicants, you're essentially treating all students as equivalent (and matching all of them).

Once the likelihood is computed, matriculation_probability estimates the probability that the given applicant will accept an offer. select_applicant allows you to randomly sample these by likelihood, and serves as the basis for running simulations about outcomes.

API reference

AdmissionsSimulation.NormalizedApplicant — Type

NormalizedApplicant holds normalized data about an applicant who received, or may receive, an offer of admission.

program::String

The abbreviation of the program the applicant was admitted to. AdmissionsSimulation.program_lookups contains the list of valid choices, together with their full names.

season::Int16

The year in which the applicant's decision was due. E.g., if the last date was April 15th, 2021, this would be 2021.

normrank::Union{Missing, Float32}

Normalized rank of the applicant: the top applicant has a rank near 0 (e.g., 1/302), and the bottom applicant has rank 1. The rank is computed among all applicants, not just those who received an offer of admission.

normofferdate::Float32

Normalized date at which the applicant received the offer of admission. 0 = date of first offer of season, 1 = decision date (typically April 15th). Candidates who were admitted in the first round would have a value of 0 (or near it), whereas candidates who were on the wait list and eventually received offers would have a larger value for this parameter.

normdecidedate::Union{Missing, Float32}

Normalized date at which the applicant replied with a decision. This uses the same scale as normofferdate. Consequently, an applicant who decided almost immediately would have a normdecidedate shortly after the normofferdate, whereas a candidate who decided on the final day will have a value of 1.0.

Use missing if the applicant has not yet decided.

accept::Union{Missing, Bool}

true if the applicant accepted our offer, false if not. Use missing if the applicant has not yet decided.

AdmissionsSimulation.NormalizedApplicant — Method

normapp = NormalizedApplicant(applicant; program_history)

Convert applicant from "natural" units to normalized units. applicant may be anything having the following fields:

program: a string encoding the program

rank::Int: the rank of the applicant compared to other applicants to the same program in the same season. Use 1 for the top candidate; the bottom candidate should have rank equal to the number of applications received.

offerdate: the date on which an offer was (or might be) extended. E.g., Date("2021-01-13").

decidedate: the date on which the candidate replied with a verdict, or missing

accept: true if the candidate accepted our offer, false if it was turned down, missing if it is unknown.

program_history should be a dictionary compatible with program_data.

AdmissionsSimulation.ProgramData — Type

ProgramData stores summary data for a particular program and admissions season.

target_raw::Int64

The target number of matriculants, based on applicant pool and training capacity.

target_corrected::Int64

The actual target, correcting for over- or under-recruitment in previous years.

nmatriculants::Union{Missing, Int64}

The number of matriculated students, or missing.

napplicants::Int64

The number of applicants received.

firstofferdate::Dates.Date

The date on which the first offer was made, essentially the beginning of the decision period for the applicants.

lastdecisiondate::Dates.Date

The date on which all applicants must have rendered a decision, or the offer expires.

AdmissionsSimulation.ProgramKey — Type

ProgramKey stores the program name and admissions season.

program::String

The program abbreviation. AdmissionsSimulation.program_lookups contains the list of valid choices, together will full names.

season::Int16

The enrollment year. This is the year in which the applicant's decision was due. E.g., if the last date was April 15th, 2021, this would be 2021.

AdmissionsSimulation.match_function — Method

fmatch = match_function(; matchprogram=false, σr=Inf32, σt=Inf32)

Generate a matching function comparing two applicants.

fmatch(template::NormalizedApplicant, applicant::NormalizedApplicant, tnow::Union{Real,Missing})

will return a number between 0 and 1, with 1 indicating a perfect match. template is the applicant you wish to find a match for, and applicant is a candidate match. tnow is used to exclude applicants who had already decided by tnow.

The parameters of fmatch are determined by criteria:

matchprogram::Bool: if true, only students from the same program are considered (all others return 0.0)

σr: the standard deviation of normrank (use Inf or missing if you don't want to consider rank in matches)

σt: the standard deviation of normofferdate (use Inf or missing if you don't want to consider offer date in matches)

AdmissionsSimulation.match_likelihood — Method

likelihood = match_likelihood(fmatch, past_applicants, applicant, tnow::Date; program_history)

Use this format if supplying tnow in Date format.

AdmissionsSimulation.match_likelihood — Method

likelihood = match_likelihood(fmatch::Function,
                              past_applicants::AbstractVector{NormalizedApplicant},
                              applicant::NormalizedApplicant,
                              tnow::Real)

Compute the likelihood among past_applicants for matching applicant. tnow is the current date in normalized form (see normdate), and is used to exclude previous applicants who had already made a decision by tnow.

See also: match_function, select_applicant.

AdmissionsSimulation.matriculation_probability — Method

p = matriculation_probability(likelihood, past_applicants)

Compute the probability that applicants weighted by likelihood would matriculate into the program, based on the choices made by past_applicants.

likelihood can be computed by match_likelihood.

AdmissionsSimulation.normdate — Method

normdate(t::Date, pdata)

Express t as a fraction of the gap between the first offer date and last decision date as stored in pdata (see program_data).

AdmissionsSimulation.program_data — Method

pdata = program_data(applicant, program_history)

Return data for the program and admission season matching applicant. One accepted format is the following:

program_history = Dict(ProgramKey(season=2021, program="NS") => ProgramData(slots=15, napplicants=302, firstofferdate=Date("2021-01-13"), lastdecisiondate=Date("2021-04-15")),
                       ProgramKey(season=2021, program="CB") => ProgramData(slots=5,  napplicants=160, firstofferdate=Date("2021-01-6"),  lastdecisiondate=Date("2021-04-15")),
)

AdmissionsSimulation.read_applicant_data — Method

past_applicants = read_applicant_data(filename; program_history)
past_applicants = read_applicant_data(f, filename; program_history)

Read past applicant data from a file. See "/home/runner/work/AdmissionsSimulation.jl/AdmissionsSimulation.jl/src/test/data/applicantdata.csv" for an example of the format.

The second form allows you to transform each row with f(row) before extracting the data. This allows you to handle input formats that differ from the default.

AdmissionsSimulation.read_program_history — Method

program_history = read_program_history(filename)
program_history = read_program_history(f, filename)

Read program history from a file. See "/home/runner/work/AdmissionsSimulation.jl/AdmissionsSimulation.jl/src/test/data/programdata.csv" for an example of the format.

The second form allows you to transform each row with f(row) before extracting the data. This allows you to handle input formats that differ from the default.

AdmissionsSimulation.run_simulation — Method

nmatriculants = run_simulation(pmatric, nsim::Int)

Given a list of candidates with probability of matriculation pmatric, perform nsim simulations of their admission decisions and compute the total number of matriculants in each simulation.

AdmissionsSimulation.select_applicant — Method

past_applicant = select_applicant(clikelihood, past_applicants)

Select a previous applicant from among past_applicants, using the cumulative likelihood clikelihood. This can be computed as cumsum(likelihood), where likelihood is computed by match_likelihood.