AdmissionsSimulation

AdmissionsSimulation is designed to simulate outcomes of graduate admissions, with the intent of helping make decisions about both initial offers and wait-list offers.

This package focuses on those students to whom offers of admission have been extended, and uses past applicants as proxies for current applicants to make predictions about whether they'll accept the offer. Because two "similar" students might end up making different final decisions, the recommended practice is to identify many different potential proxies and use the distribution of their decisions to simulate future outcomes. This is similar to a k-nearest neighbors algorithm, but with continuous weights applied to each potential neighbor and no limit on the number of neighbors used.

The inputs to this process are:

records on previous applicants

criteria for deciding the similarity between two students, resulting in a matching function computing a value between 0 (no match) and 1 (a perfect match).

The core concept may be easily explained by a simplified example. Imagine that the admissions committee for "program A" has ranked their applicants for the season. Let's focus on one "test" applicant, say their 2nd most highly-ranked applicant. Let's imagine that the history of past admissions seasons looks like the first four columns below:

Applicant	Program	Applicant rank	Accepted?	Similarity to test applicant

PastApplicant1	A	1	no	0.6

PastApplicant2	A	2	yes	1.0

PastApplicant3	A	3	no	0.6

PastApplicant4	A	4	yes	0.4

PastApplicant5	B	1	yes	0.1

PastApplicant6	B	2	no	0.15

The final column is specific to this particular "test" applicant–the one being made an offer in the current season whose response to the offer of admission is currently unknown–and must be recomputed for each current applicant we want to examine. The idea is that we look at all previous applicants who were also offered admission and assess similarity. In this table, 4 of the applicants are also from "program A." These are viewed as being especially similar to the test applicant, particularly the applicant who was also the 2nd most highly-ranked applicant in zir year (with a similarity score of "1.0"). However, in this case our matching function also allows us to use "intelligence" from other programs, and so two applicants to "program B" have a small but nonzero similarity to the test applicant.

Now, to make predictions about how our test student will respond to our offer, we form a similarity-weighted sum of the previous yes/no decisions. In this case, the probability of "yes" is (1.0 + 0.4 + 0.1)/(0.6 + 1.0 + 0.6 + 0.4 + 0.1 + 0.15) ≈ 0.53. This gives us a quantitative assessment of the likelihood that our test applicant will accept our offer. While each individual applicant must ultimately answer either yes or no, the fundamental thesis of this package is that having a per applicant matriculation probability can allow better control over class sizes through management of offer-extensions and wait-lists.

The actual model incorporates more factors than shown in this simplified example. For example, we track the date on which each past applicant informed us of zir decision. This allows us to account for the ways in which both applicant and program competitiveness may interact to lead some applicants to respond immediately to our offer of admission and others to wait until the last day of the season (typically April 15th), with the reasons for delay also being coupled to the likelihood of accepting the offer of admission (e.g., more "yes" decisions arrive early and more "no" decisions arrive late). Accounting for these additional factors in our similarity computation improves accuracy in the projections for managing the wait list.

The remainder of this documentation describes the actual implementation.

Applicant records

Records on previous applicants are in two forms: very general information is stored in program_history, a dictionary recording just a few bits of information for each program. For example:

    program_history = Dict(ProgramKey(season=2021, program="NS") => ProgramData(slots=15, napplicants=302, firstofferdate=Date("2021-01-13"), lastdecisiondate=Date("2021-04-15")),
                           ProgramKey(season=2021, program="CB") => ProgramData(slots=5,  napplicants=160, firstofferdate=Date("2021-01-6"),  lastdecisiondate=Date("2021-04-15")))

suffices to record aggregate data for two programs, "NS" and "CB", during the 2021 season (corresponding to a decision deadline of April 15, 2021). This records the target number of matriculants (slots), the total number of applications received, the date of the very first offers extended, and the date on which a decision was due.

Valid choices for program names are listed in AdmissionsSimulation.program_lookups; internally the code always uses the abbreviation, but it is possible to supply it in long form too.

Detailed applicant records only need to include applicants to whom an offer of admission was extended. The requirements are described by NormalizedApplicant(applicant; program_history).

You can load both the program history and data on applicants using read_program_history and read_applicant_data.

Match criteria

Applicants 1 and 2 are matched by the following function:

\[\phi(p_1, p_2) \psi(r_1-r_2, t_1-t_2)\]

$\phi(p_1, p_2)$ is a measure of similarity between the two applicants' programs, $p_1$ and $p_2$, based on program selectivity and yield. The $\psi$ term is applicant-specific, analyzing rank $r$ and date $t$ on which the offer of admission was made.

In the current implementation, programs are compared by selectivity $s$ and yield $y$. Selectivity is simply the fraction of applicants who receive offers of admission; yield is more complicated, because the proper handling of the wait list requires not just an estimate of how many applicants accept our offer, but also the typical timing with which they accept. A program that ranks first in the world (or one whose admissions committee targets a lot of high-certainty applicants to reduce their risk of receiving rejections) might imagine receiving a lot of "yes" replies as soon as offers are extended, whereas a program with a lot of competition might have a larger number of delayed responses. Consequently, by default the admissions season is broken into thirds, and the fraction of accepts/declines in each third is the basis for comparing yield in two programs. The formula for doing this is

\[\phi(p_1, p_2) = \exp\left(-\frac{(s_1 - s_2)^2}{2\sigma_\text{sel}^2} - \frac{({\bf y}_1 - {\bf y}_2)^2}{2\sigma_\text{yield}^2}\right).\]

Here, ${\bf y}$ is a vector encoding the fraction of accepts/declines in each period of the season, and a Euclidean distance is computed.

The $\sigma$ parameters measure the standard deviation, i.e., the tolerance for mismatch (larger $\sigma$ are more tolerant of mismatch). In the extreme of $\sigma \ll 1$, each program matches only itself; in the extreme $\sigma \gg 1$, each program matches every other program perfectly. In between, programs will draw more "intelligence" from other programs with similar selectivity and yield dynamics as their own. In practice, model-tuning (see below) seems to favor programs primarily relying on their own history.

The remaining terms are applicant-, rather than program-, specific:

\[\psi(r_1-r_2, t_1-t_2) = \exp\left( - \frac{(r_1 - r_2)^2}{2 \sigma_r^2} - \frac{(t_1 - t_2)^2}{2 \sigma_t^2}\right)\]

$r$ refers to the normalized ranks and $t$ to the normalized offer date. Smaller $\sigma_r$ increases the importance of choosing applicants of similar rank, and would indicate that there may be a strong rank-dependent element to recruitment (e.g., "more competitive applicants are harder to recruit"). Smaller $\sigma_t$ increases the importance of the timing of the offer, and would indicate that wait-list offers should be treated quite differently from initial offers. The units of both $\sigma$ parameters are those of the NormalizedApplicants, i.e., with values ranging between 0 and 1. For example, setting $\sigma_t = 0.2$ would, in essence, break the decision period (e.g., mid-January to April 15th) into roughly 5 periods, and match primarily against applicants who were extended offers during the same period; conversely, setting $\sigma_t = 5$ would mean that the timing of the offer is essentially irrelevant to predicting the decision. Note that the applicant ranks will typically be quite heavily weighted to low values (e.g., a program that only accepts the top 20% of applicants will only exhibit ranks between 0.0 and 0.2), and as a consequence $\sigma_r$ must be smaller than this span to have large effect.

One crucial point is that the matching function returns zero when comparing against past applicants who had already returned their decision by this point in the admissions season. This models the currently-undecided applicants solely in terms of prior applicants who were also undecided at this point. See match_function for specific details.

How does one tune these parameters? The core idea is to train them based on past admissions seasons: if you have records extending back several years, you make matriculation predictions for year $y$ based on data from all years up to and including year $y-1$, and then compute a correlation with the actual outcome. See match_correlation for details.

Analysis and simulations

Once you've entered student records and defined a matching function, then for each outstanding offer you can compute the match likelihood of previous applicants using match_likelihood. The sum of the returned list is a rough measure of the "number" of prior applicants deemed to be a good match for the applicant you are modeling; if this value is small, your criteria for matching may be too stringent. Conversely, if this value is approximately equal to the total number of prior applicants, you're essentially treating all students as equivalent (and matching all of them).

Once the likelihood is computed, matriculation_probability estimates the probability that the given applicant will accept an offer. select_applicant allows you to randomly sample these by likelihood, and may serve as the basis for running simulations about outcomes, although run_simulation (which just uses matriculation probability) may be a more useful approach.

API reference

AdmissionsSimulation.FacultyRecord — Type

FacultyRecord stores the program affiliations and service contributions of a particular faculty member.

start::Dates.Date

Date on which the faculty member was approved to train students in any program

programs::Vector{String}

Program affiliations, in decreasing order of importance (e.g., primary, secondary, ...)

service::Vector{Pair{String, Service}}

program=>Service contributions (to any program, not just those listed in programs)

AdmissionsSimulation.NormalizedApplicant — Type

NormalizedApplicant holds normalized data about an applicant who received, or may receive, an offer of admission.

program::String

The abbreviation of the program the applicant was admitted to. AdmissionsSimulation.program_lookups contains the list of valid choices, together with their full names.

season::Int16

The year in which the applicant's decision was due. E.g., if the last date was April 15th, 2021, this would be 2021.

normrank::Union{Missing, Float32}

Normalized rank of the applicant: the top applicant has a rank near 0 (e.g., 1/302), and the bottom applicant has rank 1. The rank is computed among all applicants, not just those who received an offer of admission.

normofferdate::Float32

Normalized date at which the applicant received the offer of admission. 0 = date of first offer of season, 1 = decision date (typically April 15th). Candidates who were admitted in the first round would have a value of 0 (or near it), whereas candidates who were on the wait list and eventually received offers would have a larger value for this parameter.

normdecidedate::Union{Missing, Float32}

Normalized date at which the applicant replied with a decision. This uses the same scale as normofferdate. Consequently, an applicant who decided almost immediately would have a normdecidedate shortly after the normofferdate, whereas a candidate who decided on the final day will have a value of 1.0.

Use missing if the applicant has not yet decided.

accept::Union{Missing, Bool}

true if the applicant accepted our offer, false if not. Use missing if the applicant has not yet decided.

AdmissionsSimulation.NormalizedApplicant — Method

normapp = NormalizedApplicant(; program, rank=missing, offerdate, decidedate=missing, accept=missing, program_history)

Create an applicant from "natural" units, where rank is an integer and dates are expressed in Date format. Some are required (those without a default value), others are optional:

program: a string encoding the program

rank::Int: the rank of the applicant compared to other applicants to the same program in the same season. Use 1 for the top candidate; the bottom candidate should have rank equal to the number of applications received.

offerdate: the date on which an offer was (or might be) extended. E.g., Date("2021-01-13").

decidedate: the date on which the candidate replied with a verdict, or missing

accept: true if the candidate accepted our offer, false if it was turned down, missing if it is unknown.

program_history should be a dictionary mapping ProgramKeys to ProgramData.

AdmissionsSimulation.Outcome — Type

Outcome(ndeclines, naccepts)

Tally of the number of declines and accepts for offers of admission.

AdmissionsSimulation.ProgramData — Type

ProgramData stores summary data for a particular program and admissions season.

target_raw::Int64

The target number of matriculants, based on applicant pool and training capacity.

target_corrected::Int64

The actual target, correcting for over- or under-recruitment in previous years.

nmatriculants::Union{Missing, Int64}

The number of matriculated students, or missing.

napplicants::Int64

The number of applicants received.

firstofferdate::Dates.Date

The date on which the first offer was made, essentially the beginning of the decision period for the applicants.

lastdecisiondate::Dates.Date

The date on which all applicants must have rendered a decision, or the offer expires.

AdmissionsSimulation.ProgramKey — Type

ProgramKey stores the program name and admissions season.

program::String

The program abbreviation. AdmissionsSimulation.program_lookups contains the list of valid choices, together will full names.

season::Int16

The enrollment year. This is the year in which the applicant's decision was due. E.g., if the last date was April 15th, 2021, this would be 2021.

AdmissionsSimulation.ProgramYieldPrediction — Type

ProgramYieldPrediction records mid-season predictions and data for a particular program.

nmatriculants::Measurements.Measurement{Float32}

The predicted number of matriculants.

priority::Float32

The program's priority for receiving wait list offers. The program with the highest priority should get the next offer. Priority is computed as deficit/stddev, where deficit is the predicted undershoot (which might be negative if the program is predicted to overshoot) and stddev is the square root of the target (Poisson noise). Thus, programs are prioritized by the significance of the deficit.

poutcome::Union{Missing, Float32}

The two-tailed p-value of the actual outcome (if supplied). This includes the effects of any future wait-list offers.

AdmissionsSimulation.Service — Type

Service measures the service contributions of a particular faculty member to a particular program.

ninterviews::Int64

The number of admissions interviews conducted for that program, typically over a fixed time window.

ncommittees::Int64

The number of thesis committees served on for that program, typically over a fixed time window.

AdmissionsSimulation.ServiceCalibration — Type

ServiceCalibration is meant to standardize forms of service to account for the fact that young programs may not have students in thesis committees and therefore don't have as much service per faculty member.

AdmissionsSimulation.aggregate — Method

proghist = aggregate(program_history::ListPairs{ProgramKey, ProgramData}, mergepairs)

Aggregate program history, merging program from => to pairs from mergepairs.

AdmissionsSimulation.aggregate — Method

facrecsnew = aggregate(facrecs::ListPairs{String,FacultyRecord}, mergepairs)

Aggregate faculty records, merging program from => to pairs from mergepairs.

AdmissionsSimulation.cached_similarity — Method

fsim = cached_similarity(σsel, σyield; offerdata, yielddata)

Cache the result of program_similarity, creating a function fsim(program1::AbstractString, program2::AbstractString) to compute the similarity between program1 and program2.

AdmissionsSimulation.calibrate_service — Function

sc = calibrate_service(progsvc, yrthresh = <7 years ago today>)

Calculate an equivalence between different forms of service. This is to handle the fact that young programs don't provide opportunities for service in the form of thesis commmittees. progsvc is from program-service, and yrthresh selects "old" programs (ones that existed prior to yrthresh) useful for calibration.

sc allows a calculation of total service based on the maximum score computed from interviews or from committees.

AdmissionsSimulation.faculty_affiliations — Function

naffil = faculty_affiliations(facrecs, scheme)

Compute the number of affiliations per program. facrecs is a list of facultyname::String => facrec::FacultyRecord pairs containing details about the affiliations of each faculty member (see FacultyRecord). facrecs can be read via read_faculty_data.

scheme controls the weighting of affiliations for faculty members with more than one affiliation:

:primary: count only the faculty member's primary affiliation (one vote/faculty)

:all: count all affiliations (multiple votes/faculty depending on the number of affiliations)

:normalized: one vote per faculty, spread equally among all that faculty member's affiliations

:weighted: one vote per faculty, with decreasing weight. For a faculty member with 3 affiliations, they would be assigned a ratio of 3 to 2 to 1. Hence the primary program would get 3/6=0.5, secondary 2/6=0.33, and tertiary 1/6=0.17.

AdmissionsSimulation.faculty_effort — Method

faculty, programs, E = faculty_effort(facrecs, daterange::AbstractRange; sc=nothing, progrange=<default>)

Use facrecs, a list of facultyname::String => facrec::FacultyRecord pairs, to estimate the average annual effort (in hours) contributed in the forms of service tracked in Service. See FacultyRecord and read_faculty_data. daterange specifies the time span covered by facrecs; it could be a Date range or something like 2016:2020 to indicate a span in calendar years. The optional sc allows you to supply a service calibration, see calibrate_service. progyears lets you supply a Dict(progname => yearrange) specifying the duration of existence of each program; the default effectively assumes you've called aggregate on facrecs to consolidate defunct programs into their modern equivalents.

On output, E is a length(faculty)-by-length(programs) matrix, where E[j,i] measures the average annual effort committed by faculty member faculty[j] to programs[i].

AdmissionsSimulation.faculty_involvement — Method

f = faculty_involvement(E::AbstractMatrix; scheme=:normeffort, annualthresh=2, M=size(E,1))

Compute the effective number of faculty f[i] involved in program i, based on annual effort E as computed by faculty_effort. annualthresh is the number of hours that must be exceeded in order to qualify as a contributing faculty member.

There are three schemes available:

:thresheffort: f[i] gets a +1 contribution from faculty member j if j exceeded annualthresh in that program. (This allows one faculty member to count multiple times.)

:normeffort (the default): for each faculty member who's total service hours across all programs exceeds annualthresh, distribute a total of one vote in proportion to service per program.

:effortshare: calculate the average service per faculty member (M faculty members total) for each program. If a faculty member exceeded this threshold for k programs, add 1/k to each. annualthresh plays no role here.

AdmissionsSimulation.match_correlation — Method

match_correlation(σsels::AbstractVector, σyields::AbstractVector, σrs::AbstractVector, σts::AbstractVector;
                  applicants, program_history, kwargs...)

Compute the prediction accuracy using historical data. For each year in program_history other than the earliest, use prior data to predict the probability of matriculation for each applicant.

The σ lists contain the values that will be used to compute accuracy; the return value is a 4-dimensional array evaluating the correlation between estimated matriculation probability and acceptance for all possible combinations of these parameters. σsel and σyield will be used by cached_similarity to determine program similarity; σr and σs will be used to measure applicant similarity.

Tuning essentially corresponds to picking the index of the entry of the return value and then setting each parameter accordingly:

corarray = match_correlation(σsels, σyields, σrs, σts; applicants, program_history)
idx = argmax(corarray)
σsel, σyield, σr, σt = σsels[idx[1]], σyields[idx[2]], σrs[idx[3]], σts[idx[4]]

AdmissionsSimulation.match_correlation — Method

match_correlation(σsel::Real, σyield::Real, σr::Real, σt::Real;
                  applicants, past_applicants, offerdata, yielddata,
                  ptail=0.0f0, minfrac=0.0)

Compute the correlation between estimated matriculation probability and decline/accept for a list of applicants' matriculation decisions. This function is used to evaluate the accuracy of predictions made by specific model parameters.

The σ arguments are matching parameters, see program_similarity and match_function. offerdata and yielddata are computed by functions of the same name. ptail is used to clamp the estimated matriculation probability between bounds, clamp(pmatric, ptail, 1-ptail). minfrac expresses the minimum fraction of past_applicants allowed to be matched; any test_applicant matching fewer than these (in the sense of the sum of likelihoods computed by match_likelihood) leads to a return value of NaN.

AdmissionsSimulation.match_function — Method

fmatch = match_function(; σr=Inf32, σt=Inf32, progsim=default_similarity)

Generate a matching function comparing two applicants.

fmatch(template::NormalizedApplicant, applicant::NormalizedApplicant, tnow::Union{Real,Missing})

will return a number between 0 and 1, with 1 indicating a perfect match. template is the applicant you wish to find a match for, and applicant is a candidate match. tnow is used to exclude applicants who had already decided by tnow.

The parameters of fmatch are determined by criteria:

σr: the standard deviation of normrank (use Inf or missing if you don't want to consider rank in matches)

σt: the standard deviation of normofferdate (use Inf or missing if you don't want to consider offer date in matches)

progsim: a function progsim(program1, program2) computing the "similarity" between programs. See cached_similarity. The default returns true if program1 == program2 and false otherwise.

AdmissionsSimulation.match_likelihood — Method

likelihood = match_likelihood(fmatch, past_applicants, applicant, tnow::Date; program_history)

Use this format if supplying tnow in Date format.

AdmissionsSimulation.match_likelihood — Method

likelihood = match_likelihood(fmatch::Function,
                              past_applicants::AbstractVector{NormalizedApplicant},
                              applicant::NormalizedApplicant,
                              tnow::Real)

Compute the likelihood among past_applicants for matching applicant. tnow is the current date in normalized form (see normdate), and is used to exclude previous applicants who had already made a decision by tnow.

See also: match_function, select_applicant.

AdmissionsSimulation.matriculation_probability — Method

p = matriculation_probability(likelihood, past_applicants)

Compute the probability that applicants weighted by likelihood would matriculate into the program, based on the choices made by past_applicants.

likelihood can be computed by match_likelihood.

AdmissionsSimulation.normdate — Method

normdate(t::Date, pdata::ProgramData)

Express t as a fraction of the gap between the first offer date and last decision date as stored in pdata (see ProgramData).

AdmissionsSimulation.offerdata — Method

offerdata(applicants, program_history)

Summarize application and offer data for each program. The output is a dictionary mapping programname => (noffers, napplicants). The program selectivity is the ratio noffers/napplicants.

AdmissionsSimulation.program_service — Method

progsvc = program_service(facrecs)

Compute the total service for each program. progsvc is a Dict(progname => ::Service).

AdmissionsSimulation.program_similarity — Method

program_similarity(program1::AbstractString, program2::AbstractString;
                   σsel=Inf32, σyield=Inf32, offerdata, yielddata)

Compute the similarity between program1 and program2, based on selectivity (fraction of applicants who are admitted) and yield (fraction of offers that get accepted). The similarity ranges between 0 and 1, with 1 corresponding to identical programs.

The keyword arguments are the parameters controlling the similarity computation. offerdata and yielddata are the outputs of two functions of the same name (offerdata and yielddata). σsel and σyield are the standard deviations of selectivity and yield. The similarity is computed as

\[\exp\left(-\frac{(s₁ - s₂)²}{2σ_\text{sel}²} - \frac{(y₁ - y₂)²}{2σ_\text{yield}²}\right).\]

The output of this function can be cached with cached_similarity.

AdmissionsSimulation.read_applicant_data — Method

past_applicants = read_applicant_data(filename; program_history)

Read past applicant data from a file. See "/home/runner/work/AdmissionsSimulation.jl/AdmissionsSimulation.jl/src/test/data/applicantdata.csv" for an example of the format.

AdmissionsSimulation.read_faculty_data — Method

facrecs = read_faculty_data(filename)

Read faculty training participation from a file. See "/home/runner/work/AdmissionsSimulation.jl/AdmissionsSimulation.jl/src/test/data/facultyinvolvement.csv" for an example of the format.

AdmissionsSimulation.read_program_history — Method

program_history = read_program_history(filename)

Read program history from a file. See "/home/runner/work/AdmissionsSimulation.jl/AdmissionsSimulation.jl/src/test/data/programdata.csv" for an example of the format.

AdmissionsSimulation.run_simulation — Function

nmatriculants = run_simulation(pmatrics::AbstractVector, nsim::Int=100)

Given a list of candidates each with probability of matriculation pmatrics[i], perform nsim simulations of their admission decisions and compute the total number of matriculants in each simulation.

AdmissionsSimulation.select_applicant — Method

past_applicant = select_applicant(clikelihood, past_applicants)

Select a previous applicant from among past_applicants, using the cumulative likelihood clikelihood. This can be computed as cumsum(likelihood), where likelihood is computed by match_likelihood.

AdmissionsSimulation.targets — Method

targets(program_applicants, fiis, N)

Compute the target number of matriculants for each program. program_applicants is a collection of program => napplicants pairs; fiis is a collection of program => FII scores (see faculty_involvement). N is the total number of matriculants across all programs.

Each program gets weighted by the geometric mean of the # of applicants and FII.

AdmissionsSimulation.wait_list_analysis — Method

nmatric, progstatus = wait_list_analysis(fmatch::Function,
                                         past_applicants::AbstractVector{NormalizedApplicant},
                                         applicants::AbstractVector{NormalizedApplicant},
                                         tnow::Date;
                                         program_history,
                                         actual_yield=nothing)

Compute the estimated number nmatric of matriculants and the program-specific yield prediction and wait-list priority, progstatus. progstatus is a mapping progname => progyp::ProgramYieldPrediction (see ProgramYieldPrediction).

The arguments are similarly to match_likelihood. If you're doing a post-hoc analysis, actual_yield can be a Dict(progname => nmatric), in which case the p-value for the observed outcome will also be stored in progstatus.

AdmissionsSimulation.yielddata — Method

yielddata(Outcome, applicants)
yielddata(Tuple{Outcome,Outcome,Outcome}, applicants)

Compute the outcome of offers of admission for each program. applicants should be a list of NormalizedApplicant. The first form computes the Outcome for the entire season, and the second breaks the season up into epochs of equal duration and computes the outcome for each epoch separately. If provided, program_similarity will make use of the time-dependence of the yield.