API reference: Admit

Admit.NormalizedApplicant — Type

NormalizedApplicant holds normalized data about an applicant who received, or may receive, an offer of admission.

applicantdata::PersonalData

Individual data about the applicant, see PersonalData.

program::String

The abbreviation of the program the applicant was admitted to. Admit.program_lookups contains the list of valid choices, together with their full names.

season::Int16

The year in which the applicant's decision was due. E.g., if the decision deadline was April 15th, 2021, this would be 2021.

normrank::Union{Missing, Float32}

Normalized rank of the applicant: the top applicant has a rank near 0 (e.g., 1/302), and the bottom applicant has rank 1. The rank is computed among all applicants, not just those who received an offer of admission.

normofferdate::Union{Missing, Float32}

Normalized date at which the applicant received the offer of admission. 0 = date of first offer of season, 1 = decision date (typically April 15th). Candidates who were admitted in the first round would have a value of 0 (or near it), whereas candidates who were on the wait list and eventually received offers would have a larger value for this parameter.

normdecidedate::Union{Missing, Float32}

Normalized date at which the applicant replied with a decision. This uses the same scale as normofferdate. Consequently, an applicant who decided almost immediately would have a normdecidedate shortly after the normofferdate, whereas a candidate who decided on the final day will have a value of 1.0.

Use missing if the applicant has not yet decided.

accept::Union{Missing, Bool}

true if the applicant accepted our offer, false if not. Use missing if the applicant has not yet decided.

Admit.NormalizedApplicant — Method

normapp = NormalizedApplicant(; program, urmdd=missing, foreign=missing, rank=missing, offerdate, decidedate=missing, accept=missing, program_history)

Create an applicant from "natural" units, where rank is an integer and dates are expressed in Date format. Some are required (those without a default value), others are optional:

program: a string encoding the program

urmdd: true if applicant is a URM, disadvantaged, or disabled

foreign: true if applicant is not a citizen or permanent resident

rank::Int: the rank of the applicant compared to other applicants to the same program in the same season. Use 1 for the top candidate; the bottom candidate should have rank equal to the number of applications received.

offerdate: the date on which an offer was (or might be) extended. E.g., Date("2021-01-13").

decidedate: the date on which the candidate replied with a verdict, or missing

accept: true if the candidate accepted our offer, false if it was turned down, missing if it is unknown.

program_history should be a dictionary mapping ProgramKeys to ProgramData.

Admit.Outcome — Type

Outcome(ndeclines, naccepts)

Tally of the number of declines and accepts for offers of admission.

Admit.PersonalData — Type

PersonalData holds relevant data about an individual applicant.

name::String

Name of the applicant

urmdd::Union{Missing, Bool}

true if an applicant is an underrepresented minority, disadvantaged, or disabled.

foreign::Union{Missing, Bool}

true if an applicant is not a US citizen or permanent resident.

Admit.ProgramData — Type

ProgramData stores summary data for a particular program and admissions season.

target_raw::Int64

The target number of matriculants, based on applicant pool and training capacity.

target_corrected::Int64

The actual target, correcting for over- or under-recruitment in previous years.

nmatriculants::Union{Missing, Int64}

The number of matriculated students, or missing.

napplicants::Int64

The number of applicants received.

firstofferdate::Dates.Date

The date on which the first offer was made, essentially the beginning of the decision period for the applicants.

lastdecisiondate::Dates.Date

The date on which all applicants must have rendered a decision, or the offer expires.

Admit.ProgramKey — Type

ProgramKey stores the program name and admissions season.

program::String

The program abbreviation. Admit.program_lookups contains the list of valid choices, together will full names.

season::Int16

The enrollment year. This is the year in which the applicant's decision was due. E.g., if the last date was April 15th, 2021, this would be 2021.

Admit.ProgramYieldPrediction — Type

ProgramYieldPrediction records mid-season predictions and data for a particular program.

nmatriculants::Measurements.Measurement{Float32}

The predicted number of matriculants.

priority::Float32

The program's priority for receiving wait list offers. The program with the highest priority should get the next offer. Priority is computed as deficit/stddev, where deficit is the predicted undershoot (which might be negative if the program is predicted to overshoot) and stddev is the square root of the target (Poisson noise). Thus, programs are prioritized by the significance of the deficit.

poutcome::Union{Missing, Float32}

The two-tailed p-value of the actual outcome (if supplied). This includes the effects of any future wait-list offers.

Admit.add_offers! — Function

nmatric, pq0=>pq = add_offers!(fmatch, program_offers::Dict, program_candidates::Dict, past_applicants, tnow::Date=today(), σthresh=2; program_history)

Transfer applicants from program_candidates to program_offers depending on whether projections in nmatric are below DBBS-wide target by at least σthresh standard deviations. nmatric0 is computed upon entrance, while nmatric reflects updated projections after adding offers. Likewise pq0 is the program-priority initially, and pq after adding offers.

tnow should be the date for which the computation should be performed, and is used to determine whether candidates have already informed us of their decision.

Programs are prioritized for offers by the deficit divided by the expected noise.

Admit.aggregate — Method

proghist = aggregate(program_history::ListPairs{ProgramKey, ProgramData}, mergepairs)

Aggregate program history, merging program from => to pairs from mergepairs.

Admit.cached_similarity — Method

fsim = cached_similarity(σsel, σyield; offerdata, yielddata)

Cache the result of program_similarity, creating a function fsim(program1::AbstractString, program2::AbstractString) to compute the similarity between program1 and program2.

Admit.generate_fake_candidates — Function

program_candidates = generate_fake_candidates(program_history, season::Integer, program_offer_dates=nothing)

Generate fake candidates for each program, each ranked starting from 1. season is the year for which the offers should be generated. Optionally provide program_offer_dates, a program_name=>list_of_offer_dates dictionary which will be used to generate offer dates (by default set to the firstofferdate in program_history). Unless σt is quite small in the matching function, adding additional offer dates (e.g., for multiple interview dates) may not change the outcome substantially.

On output, program_candidates is a Dict(program1=>[applicant1a, applicant1b, ...], ...) storing the fake applicants per program in rank order. See initial_offers!.

Example

This calculates the number of initial offers per program:

julia> program_candidates = Admit.generate_fake_candidates(program_history, 2021);

julia> program_offers = initial_offers!(fmatch, program_candidates, past_applicants, Date("2021-01-01"); program_history);

julia> noffers = sort([prog => length(list) for (prog, list) in program_offers]; by=first)
13-element Vector{Pair{String, Int64}}:
  "BBSB" => 16
  "BIDS" => 9
    "CB" => 13
   "CSB" => 14
 "DRSCB" => 14
  "EEPB" => 12
   "HSG" => 9
   "IMM" => 16
   "MCB" => 16
   "MGG" => 18
  "MMMP" => 18
    "NS" => 31
   "PMB" => 14

Admit.initial_offers! — Method

program_offers = initial_offers!(fmatch, program_candidates::Dict, past_applicants, tnow::Date=today(), σthresh=2; program_history)

Allocate initial offers of admission at the beginning of the season. See add_offers! for more information. See also generate_fake_candidates to plan offers in cases where some programs want to make their initial offers before other programs have finished interviewing.

Admit.manage_offers — Function

app = manage_offers(fetch_past_applicants::Function, fetch_applicants::Function, fetch_program_data::Function, tnow::Union{Date,Function}=today, σthresh::Real=2;
                    σsel=0.2f0, σyield=1.0f0, σr=0.5f0, σt=Inf32)

Create a report about the current state of admissions. The report has 3 tabs:

"Summary" gives an overview across all programs, along with a list of candidates deemed ready to receive an offer given the target in the entry box at the top right.

"Program" provides a detailed view of a single program selected in the dropdown at the top right, showing candidates who have accepted or rejected the offer of admission, along with a predicted matriculation probability for each undecided applicant.

"Internals" provides information about the model details, currently just the "Program similarity" score which expresses the use of cross-program data in predicting matriculation probability.

The "Refresh applicants" button fetches fresh data about the current applicants, and should be used to update projections as decisions get made.

The input arguments are:

fetch_past_applicants() should return a list of NormalizedApplicants from previous admissions years. These will be used to make predictions about the decisions of current applicants.

fetch_applicants() should return a list of NormalizedApplicants from the current season. This gets called every time you click "Refresh applicants".

fetch_program_data() should return a Dict{ProgramKey,ProgramData} containing overall statistics about the program over the time covered by the combination of fetch_past_applicants and fetch_applicants.

tnow can either be a Date or a function returning a Date. The default, the function Dates.today, will update the date every time the page renders. Pass a static date only if you don't want the date updating.

σthresh determines how conservative the system will be in extending wait-list offers. On the "Summary" tab, the "Total estimate" will be expressed as mean ± stddev, and mean + σthresh * stddev will be held approximately at the total target. Larger values of σthresh make it less likely that you'll overshoot, and more likely that you'll undershoot. Values in the range of 0-3 seem like plausible choices (default: 2).

σsel and σyield determine the cross-program similarity (see program_similarity)

σr and σt determine how much the prediction uses applicant rank and offer date, respectively (see match_function).

To render the report in a browser, use

Admit.run_server(app, "0.0.0.0", debug=true)

Admit.match_correlation — Method

match_correlation(matchcreator::Function, σlists::AbstractVector...;
                  applicants, program_history, kwargs...)

Compute the prediction accuracy using historical data. For each year in program_history other than the earliest, use prior data to predict the probability of matriculation for each applicant.

The σ lists contain the values that will be used to compute accuracy; the return value is an n-dimensional array evaluating the correlation between estimated matriculation probability and acceptance for all possible combinations of these parameters. matchcreator(σ1, σ2...; offerdata, yielddata) should return a similarity-computing function fmatch(template, applicant, tnow) using the specific σs provided.

Tuning essentially corresponds to picking the index of the entry of the return value and then setting each parameter accordingly:

corarray = match_correlation(Admit.fmatch_prog_rank_date, σsels, σyields, σrs, σts; applicants, program_history)
idx = argmax(corarray)
σsel, σyield, σr, σt = σsels[idx[1]], σyields[idx[2]], σrs[idx[3]], σts[idx[4]]

fmatch_prog_rank_date is a default and can be omitted if you want to use this function.

Admit.match_correlation — Method

match_correlation(fmatch; applicants, past_applicants, ptail=0.0f0, minfrac=0.01)

Compute the correlation between estimated matriculation probability and decline/accept for a list of applicants' matriculation decisions. This function is used to evaluate the accuracy of predictions made by specific model parameters.

fmatch(reference, applicant, tnow) is the similarity-computation function. ptail is used to clamp the estimated matriculation probability between bounds, clamp(pmatric, ptail, 1-ptail). minfrac expresses the minimum fraction of past_applicants allowed to be matched; any test_applicant matching fewer than these (in the sense of the sum of likelihoods computed by match_likelihood) leads to a return value of NaN.

Admit.match_function — Method

fmatch = match_function(; σr=Inf32, σt=Inf32, progsim=default_similarity)

Generate a matching function comparing two applicants.

fmatch(template::NormalizedApplicant, applicant::NormalizedApplicant, tnow::Union{Real,Missing})

will return a number between 0 and 1, with 1 indicating a perfect match. template is the applicant you wish to find a match for, and applicant is a candidate match. tnow is used to exclude applicants who had already decided by tnow.

The parameters of fmatch are determined by criteria:

σr: the standard deviation of normrank (use Inf or missing if you don't want to consider rank in matches)

σt: the standard deviation of normofferdate (use Inf or missing if you don't want to consider offer date in matches)

progsim: a function progsim(program1, program2) computing the "similarity" between programs. See cached_similarity. The default returns true if program1 == program2 and false otherwise.

Admit.match_likelihood — Method

likelihood = match_likelihood(fmatch, past_applicants, applicant, tnow::Date; program_history)

Use this format if supplying tnow in Date format.

Admit.match_likelihood — Method

likelihood = match_likelihood(fmatch::Function,
                              past_applicants::AbstractVector{NormalizedApplicant},
                              applicant::NormalizedApplicant,
                              tnow::Real)

Compute the likelihood among past_applicants for matching applicant. tnow is the current date in normalized form (see normdate), and is used to exclude previous applicants who had already made a decision by tnow.

See also: match_function, select_applicant.

Admit.matriculation_probability — Method

p = matriculation_probability(likelihood, past_applicants)

Compute the probability that applicants weighted by likelihood would matriculate into the program, based on the choices made by past_applicants.

likelihood can be computed by match_likelihood.

Admit.normdate — Method

normdate(t::Date, pdata::ProgramData)

Express t as a fraction of the gap between the first offer date and last decision date as stored in pdata (see ProgramData).

Examples

julia> using Dates

julia> pd = ProgramData(firstofferdate=Date("2021-02-11"), lastdecisiondate=Date("2021-04-15"))
ProgramData(0, 0, 0, 0, Date("2021-02-11"), Date("2021-04-15"))

julia> normdate(Date("2021-02-11"), pd)
0.0f0

julia> normdate(Date("2021-04-15"), pd)
1.0f0

julia> normdate(Date("2021-03-15"), pd)
0.50793654f0

julia> normdate(Date("2021-01-01"), pd)    # dates prior to the first offer date are negatve
-0.6507937f0

Admit.offerdata — Method

offerdata(applicants, program_history)

Summarize application and offer data for each program. The output is a dictionary mapping programname => (noffers, napplicants). The program selectivity is the ratio noffers/napplicants.

Admit.program_similarity — Method

program_similarity(program1::AbstractString, program2::AbstractString;
                   σsel=Inf32, σyield=Inf32, offerdata, yielddata)

Compute the similarity between program1 and program2, based on selectivity (fraction of applicants who are admitted) and yield (fraction of offers that get accepted). The similarity ranges between 0 and 1, with 1 corresponding to identical programs.

The keyword arguments are the parameters controlling the similarity computation. offerdata and yielddata are the outputs of two functions of the same name (offerdata and yielddata). σsel and σyield are the standard deviations of selectivity and yield. The similarity is computed as

\[\exp\left(-\frac{(s₁ - s₂)²}{2σ_\text{sel}²} - \frac{(y₁ - y₂)²}{2σ_\text{yield}²}\right).\]

The output of this function can be cached with cached_similarity.

Admit.read_applicant_data — Method

past_applicants = read_applicant_data(filename; program_history)

Read past applicant data from a file. See "/home/runner/work/AdmissionSuite.jl/AdmissionSuite.jl/Admit/src/test/data/applicantdata.csv" for an example of the format.

Admit.read_program_history — Method

program_history = read_program_history(filename)

Read program history from a file. See "/home/runner/work/AdmissionSuite.jl/AdmissionSuite.jl/Admit/src/test/data/programdata.csv" for an example of the format.

Admit.run_simulation — Function

nmatriculants = run_simulation(pmatrics::AbstractVector, nsim::Int=100)

Given a list of candidates each with probability of matriculation pmatrics[i], perform nsim simulations of their admission decisions and compute the total number of matriculants in each simulation.

Admit.select_applicant — Method

past_applicant = select_applicant(clikelihood, past_applicants)

Select a previous applicant from among past_applicants, using the cumulative likelihood clikelihood. This can be computed as cumsum(likelihood), where likelihood is computed by match_likelihood.

Admit.wait_list_analysis — Method

nmatric, progstatus = wait_list_analysis(fmatch::Function,
                                         past_applicants::AbstractVector{NormalizedApplicant},
                                         applicants::AbstractVector{NormalizedApplicant},
                                         tnow::Date;
                                         program_history,
                                         actual_yield=nothing)

Compute the estimated number nmatric of matriculants and the program-specific yield prediction and wait-list priority, progstatus. progstatus is a mapping progname => progyp::ProgramYieldPrediction (see ProgramYieldPrediction). nmatric assumes all applicants get offers.

The arguments are similarly to match_likelihood. If you're doing a post-hoc analysis, actual_yield can be a Dict(progname => nmatric), in which case the p-value for the observed outcome will also be stored in progstatus.

Admit.yielddata — Method

yielddata(Outcome, applicants)
yielddata(Tuple{Outcome,Outcome,Outcome}, applicants)

Compute the outcome of offers of admission for each program. applicants should be a list of NormalizedApplicant. The first form computes the Outcome for the entire season, and the second breaks the season up into epochs of equal duration and computes the outcome for each epoch separately. If provided, program_similarity will make use of the time-dependence of the yield.

API reference: AdmitConfiguration

AdmitConfiguration.addprogram — Method

addprogram(abbrv::AbstractString)

Dynamically add a new program with abbreviation abbrv.

This is primarily used for writing tests; most users should use setprograms instead.

AdmitConfiguration.delprogram — Method

delprogram(abbrv::AbstractString)

Dynamically delete the program with abbreviation abbrv.

This is primarily used for writing tests; most users should use setprograms instead.

AdmitConfiguration.setprograms — Function

setprograms(filename; force=false)

Configure your local programs, given a CSV file filename of the format in examples/WashU.csv. (You can create such files with a spreadsheet program, exporting the table in "Comma separated value" format.) The minimum requirement is the Abbreviation column, which must contain the list of all programs in "short-form" name. Use ProgramName if you want a long-form name, especially if you use a SQL database which sometimes uses that name.

SeasonStart, SeasonEnd, MergeTo, and SplitFrom can be used to track changes in your programs over time; this can be relevant because historical data is used to forecast matriculation probability (for Admit.jl) and faculty service (for AdmissionTargets.jl). "Season" refers to the year for final acceptance of an offer of admission, e.g., a deadline of April 15, 2013 would be season 2013. SeasonStart should contain the first season in which a program admitted applicants; SeasonEnd should contain the last season (if applicable) in which a program admitted applicants. If a defunct program merged into a newer one, set the MergeTo field as the name of the newer program; if a current program was created from a previous one, set the SplitFrom field. If a newly-created program has no clear heritage, just leave these blank (but no prior historical data will be available).

If you're re-setting the configuration from an existing one, use force=true.

Warning

You need to load the CSV package (using CSV) for setprograms to be available. See the AdmissionSuite web documentation for more information.

AdmitConfiguration.validateprogram — Method

prog = validateprogram(program::AbstractString)

Return the abbreviation prog from either a valid abbreviation or valid long-form program name program. An error is thrown if the program is not recognized.

See setprograms to configure your local programs, or addprogram and delprogram to configure them dynamically.

API reference: AdmissionTargets

AdmissionTargets.FacultyRecord — Type

FacultyRecord stores the program affiliations and service contributions of a particular faculty member.

start::Dates.Date

Date on which the faculty member was approved to train students in any program

programs::Vector{String}

Program affiliations, in decreasing order of importance (e.g., primary, secondary, ...)

service::Vector{Pair{String, Service}}

program=>Service contributions (to any program, not just those listed in programs)

AdmissionTargets.Service — Type

Service measures the service contributions of a particular faculty member to a particular program.

ninterviews::Float32

The number of admissions interviews conducted for that program, typically over a fixed time window.

ncommittees::Float32

The number of thesis committees served on for that program, typically over a fixed time window.

AdmissionTargets.ServiceCalibration — Type

ServiceCalibration is meant to standardize forms of service to account for the fact that young programs may not have students in thesis committees and therefore don't have as much service per faculty member.

AdmissionTargets.aggregate — Method

facrecsnew = aggregate(facrecs::ListPairs{<:AbstractString,FacultyRecord}, mergepairs)

Aggregate faculty records, merging program from => to pairs from mergepairs.

AdmissionTargets.calibrate_service — Function

sc = calibrate_service(progsvc, yrthresh = <7 years ago today>)

Calculate an equivalence between different forms of service. This is to handle the fact that young programs don't provide opportunities for service in the form of thesis commmittees. progsvc is from program_service, and yrthresh selects "old" programs (ones that existed prior to yrthresh) useful for calibration.

sc allows a calculation of total service based on the maximum score computed from interviews or from committees.

AdmissionTargets.faculty_affiliations — Function

naffil = faculty_affiliations(facrecs, scheme)

Compute the number of affiliations per program. facrecs is a list of facultyname::String => facrec::FacultyRecord pairs containing details about the affiliations of each faculty member (see FacultyRecord). facrecs can be read via read_faculty_data.

scheme controls the weighting of affiliations for faculty members with more than one affiliation:

:primary (default): count only the faculty member's primary affiliation (one vote/faculty)

:all: count all affiliations (multiple votes/faculty depending on the number of affiliations)

:normalized: one vote per faculty, spread equally among all that faculty member's affiliations

:weighted: one vote per faculty, with decreasing weight. For a faculty member with 3 affiliations, they would be assigned a ratio of 3 to 2 to 1. Hence the primary program would get 3/6=0.5, secondary 2/6=0.33, and tertiary 1/6=0.17.

AdmissionTargets.faculty_effort — Method

faculty, programs, E = faculty_effort(facrecs, daterange::AbstractRange; sc=nothing, progrange=<default>)

Use facrecs, a list of facultyname::String => facrec::FacultyRecord pairs, to estimate the average annual effort (in hours) contributed in the forms of service tracked in Service. See FacultyRecord and read_faculty_data. daterange specifies the time span covered by facrecs; it could be a Date range or something like 2016:2020 to indicate a span in calendar years. The optional sc allows you to supply a service calibration, see calibrate_service. progyears lets you supply a Dict(progname => yearrange) specifying the duration of existence of each program; the default effectively assumes you've called aggregate on facrecs to consolidate defunct programs into their modern equivalents.

On output, E is a length(faculty)-by-length(programs) matrix, where E[j,i] measures the average annual effort committed by faculty member faculty[j] to programs[i].

AdmissionTargets.faculty_involvement — Method

f = faculty_involvement(E::AbstractMatrix; scheme=:normeffort, annualthresh=3, M=size(E,1))

Compute the effective number of faculty f[i] involved in program i, based on annual effort E as computed by faculty_effort. annualthresh is the number of hours that must be exceeded in order to qualify as a contributing faculty member.

There are three schemes available:

:thresheffort: f[i] gets a +1 contribution from faculty member j if j exceeded annualthresh in that program. (This allows one faculty member to count multiple times.)

:normeffort (the default): for each faculty member who's total service hours across all programs exceeds annualthresh, distribute a total of one vote in proportion to service per program.

:effortshare: calculate the average service per faculty member (M faculty members total) for each program. If a faculty member exceeded this threshold for k programs, add 1/k to each. annualthresh plays no role here.

AdmissionTargets.program_service — Method

progsvc = program_service(facrecs)

Compute the total service for each program. progsvc is a Dict(progname => ::Service).

AdmissionTargets.read_faculty_data — Method

facrecs = readfacultydata(filename)

Read faculty training participation from a file. See "/home/runner/work/AdmissionSuite.jl/AdmissionSuite.jl/AdmissionTargets/src/test/data/facultyinvolvement.csv" for an example of the format.

AdmissionTargets.targets — Method

tgts, parameters = targets(program_napplicants, program_nfaculty, N, minslots)

Compute the target number of matriculants for each program, reserving some slots to ensure that no program gets fewer than minslots. Slots are assigned via a "hyperbolic" parametrization:

\[n_i = \sqrt{n_0^2 + \frac{{N^\prime}^2 w_i^2}{W^2}}.\]

n₀ and N′ are parameters computed to ensure that the minimum nᵢ (across all programs) is equal to minslots, and the sum of slots is N. wᵢ is the weight assigned to the ith program and is extracted from program_nfaculty, and W² = ∑ᵢ wᵢ^2. Very small programs might receive most of their slots from n₀, whereas large programs receive approximately a fraction N′/N < 1 of the slots they would otherwise receive without minslots.

This "hyperbolic" approach provides graded program-leveling, reserving the largest boost for small programs. It does not have sharp cutoffs: even programs that would have received barely more than minslots still get some boost.

AdmissionTargets.targets — Method

tgts = targets(program_napplicants, program_nfaculty, N)

Compute the target number of matriculants for each program. program_napplicants is a collection of program => napplicants pairs; program_nfaculty is a collection of program => nfaculty scores (see faculty_affiliations and faculty_involvement). N is the total number of matriculants across all programs.

Each program gets weighted by the geometric mean of the # of applicants and faculty. In other words, each program gets

\[n_i = N \frac{\sqrt{a_i f_i}}{\sum_k \sqrt{a_k f_k}}\]

slots. On output, tgts[program] is the number of slots for program.